requirements.txt
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
python-dotenv==1.0.1
pandas==2.2.2
twilio==9.0.5